Move `print_table_or_error` function
